Add WCS (box-style) texture coordinate mapping as WcsBox
The world coordinate system (box-style) mapping
essentially maps a coordinate to a box face on
world origo. Each face maps to the complete texture.

A position is mapped onto the face that is closest to
it.

This works best as vector input to an image texture node
that has projection set to Flat.