Wallpaper projection for EnvironmentTextureNode.
Wallpaper projection is a planar mapping on the background of
the environment texture image. The Window is used to generate
UV coordinates, similar to Window for surface shading.

This planar mapping works best when no vector input is given
from other nodes.

To facilitate the computation of the proper UV coordinates a
cameratondc transform is added to the KernelCamera and Camera
data structures.